Rename bindings to participantC and participantFortran
