+++ docs/changelog/1654.md
@@ -1,1 +1,1 @@
-- Renamed the API header files `SolverInterface.hpp` to `precice.hpp`, `SolverInterfaceC.h` to `precicec.h`, and `SolverInterfaceFortran.hpp` to `precicefortran.hpp`.
+- Renamed the API header files `SolverInterface.hpp` to `precice.hpp`, `SolverInterfaceC.h` to `preciceC.h`, and `SolverInterfaceFortran.hpp` to `preciceFortran.hpp`.
