+++ extras/bindings/c/CMakeLists.txt
@@ -1,10 +1,10 @@
 set_property(TARGET precice PROPERTY PUBLIC_HEADER
-  ${CMAKE_CURRENT_LIST_DIR}/include/precice/precicec.h
+  ${CMAKE_CURRENT_LIST_DIR}/include/precice/preciceC.h
   APPEND)
 
 target_sources(preciceCore PRIVATE
-  ${CMAKE_CURRENT_LIST_DIR}/include/precice/precicec.h
-  ${CMAKE_CURRENT_LIST_DIR}/src/precicec.cpp
+  ${CMAKE_CURRENT_LIST_DIR}/include/precice/preciceC.h
+  ${CMAKE_CURRENT_LIST_DIR}/src/preciceC.cpp
   )
 
 target_include_directories(preciceCore PUBLIC
